+++ docs/docs/about/release-notes.md
@@ -14,10 +14,14 @@ You can determine your currently installed version using `pheasant --version`:
 
 ~~~
 $ pheasant --version
-pheasant, version 2.1.3 from /path/to/pheasant (Python 3.7)
+pheasant, version 2.1.6 from /path/to/pheasant (Python 3.7)
 ~~~
 
 ## Change log
+
+### Version 2.1.6 (2019-04-08)
+
+* Add progress bar for jupyter converter.
 
 ### Version 2.1.5 (2019-04-07)